+++ lib/types/options.ts
@@ -676,18 +676,8 @@ export interface PicoCAD2ViewerState {
 export interface PicoCAD2ViewerOptions {
 	canvas?: HTMLCanvasElement;
 	context?: PicoCAD2Context;
-	shadingMode?: number;
-	renderMode?: number;
-	projectionMode?: ProjectionMode;
 	backgroundColor?: Color3 | null;
-	outlineSize?: number;
-	outlineColor?: Color3;
-	scanlines?: boolean;
-	scanlineColor?: Color3;
 	animationSpeed?: number;
-	cameraMode?: CameraMode;
-	cameraModeSpeed?: number;
-	cameraModeDirection?: "left" | "right";
 	resolution?: {
 		width: number;
 		height: number;